projects/swarm/pyproject.toml
[project]
name = "swarm"
version = "0.1.0"
description = "LangGraph-based multi-agent swarm over LiteLLM (vLLM + exo backends) with Phoenix observability and browser-use subagents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  # Orchestration
  "langgraph>=0.2.50",
  "langchain>=0.3.7",
  "langchain-openai>=0.2.10",
  # Model router — one OpenAI-compatible endpoint over all backends
  # (local vLLM, exo on :52416, and any future worker nodes).
  "litellm[proxy]>=1.52.0",
  # Observability / swarm visualization
  "arize-phoenix>=5.12.0",
  "openinference-instrumentation-langchain>=0.1.29",
  "openinference-instrumentation-openai>=0.1.18",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0",
  # Browser automation subagent
  "browser-use>=0.2.0",
  "playwright>=1.48.0",
  # Obsidian vault subagent — talks to the obsidian-mcp-server over
  # stdio. `mcp` is the official Python SDK that powers the adapter.
  "langchain-mcp-adapters>=0.2.2",
  "mcp>=1.13.0",
  # Utilities
  "httpx>=0.27.2",
  "pydantic>=2.9.2",
  "uvloop>=0.21.0",
  "typer>=0.13.0",
  "rich>=13.9.4",
  "pyyaml>=6.0.2",
  "fastapi>=0.115.5",
  "uvicorn>=0.32.1",
]

[project.scripts]
swarm = "swarm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["swarm"]

[tool.uv]
package = true
